requirements.txt
asgiref==3.5.0
astroid==2.11.2
attrs==21.4.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.4
distlib==0.3.4
Django==4.0.3
django-stubs==1.10.1
django-stubs-ext==0.4.0
djangorestframework==3.13.1
djangorestframework-stubs==1.5.0
drf-spectacular==0.22.0
filelock==3.6.0
identify==2.4.12
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.1
jsonschema==4.4.0
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mypy==0.942
mypy-extensions==0.4.3
nodeenv==1.6.0
peewee==3.14.10
platformdirs==2.5.1
pre-commit==2.18.1
pylint==2.13.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
types-Markdown==3.3.12
types-pytz==2021.3.6
types-PyYAML==6.0.5
types-requests==2.27.16
types-urllib3==1.26.11
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
wrapt==1.14.0
